Made many static elements non-static
Some variables and methods were static, which made it impossible to
store separate distinct info for the two marios. I fixed this, and also
made it so that the different marios can get their own powerups.
However, a current consequence of this (that I can probably fix) is that
both marios start the levels small instead of powered up.